First TyperController

+++ src/main/java/pl/most/typer/repository/typerrepo/TyperCompetitionRepository.java
@@ -3,6 +3,6 @@ package pl.most.typer.repository.typerrepo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.most.typer.model.typer.TyperCompetition;
 
-public interface TyperCompetitionRepository extends JpaRepository<TyperCompetition,Long> {
+public interface TyperCompetitionRepository extends JpaRepository<TyperCompetition,Integer> {
 
 }
